More reliability (and caching) in uavo tgz production
Previously, it would start streaming the tarball immediately as it began
discovering what was in it.  If an object wasn't cached and there was an
issue fetching it from origin, this couldn't be reported reliably.

Now, the entire tarball is created before the response is sent.  This
gives me the opportunity to also cache the whole response, which has the
nice property of bringing the entire response creation down to a single
memcache call in the best case instead of > 120 in the best case.